Version bump and rebranding
