--- EmbyReports.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26730.3
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmbyReports", "EmbyReports\EmbyReports.csproj", "{A2217228-D9FD-48E7-827A-B9302212FE38}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A2217228-D9FD-48E7-827A-B9302212FE38}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A2217228-D9FD-48E7-827A-B9302212FE38}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A2217228-D9FD-48E7-827A-B9302212FE38}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A2217228-D9FD-48E7-827A-B9302212FE38}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {FC91FF5C-ADA7-45F9-AB2E-B0BB1CA4BFCF}
-	EndGlobalSection
-EndGlobal
