--- EmbyReports/Api/Common/HeaderActivitiesMetadata.cs
@@ -1,16 +1,0 @@
-namespace EmbyReports.Api.Common
-{
-    public enum HeaderActivitiesMetadata
-	{
-        None,
-        Name,
-        Overview,
-        ShortOverview,
-        Type,
-        Date,
-        UserPrimaryImageTag,
-        Severity,
-        Item,
-        User
-	}
-}
