--- EmbyReports/Api/Common/ItemViewType.cs
@@ -1,21 +1,0 @@
-namespace EmbyReports.Api.Common
-{
-    public enum ItemViewType
-    {
-        None,
-        Detail,
-        Edit,
-        List,
-        ItemByNameDetails,
-        StatusImage,
-        EmbeddedImage,
-        SubtitleImage,
-        TrailersImage,
-        SpecialsImage,
-        LockDataImage,
-        TagsPrimaryImage,
-        TagsBackdropImage,
-        TagsLogoImage,
-        UserPrimaryImage
-    }
-}
