--- EmbyReports/Api/Common/ReportDisplayType.cs
@@ -1,10 +1,0 @@
-namespace EmbyReports.Api.Common
-{
-    public enum ReportDisplayType
-    {
-		None,
-		Screen,
-		Export,
-        ScreenExport 
-	}
-}
