--- EmbyReports/Api/Common/ReportExportType.cs
@@ -1,8 +1,0 @@
-namespace EmbyReports.Api.Common
-{
-    public enum ReportExportType
-	{
-		CSV,
-		Excel
-	}
-}
