--- EmbyReports/Api/Common/ReportFieldType.cs
@@ -1,15 +1,0 @@
-namespace EmbyReports.Api.Common
-{
-    public enum ReportFieldType
-	{
-		String,
-		Boolean,
-		Date,
-		Time,
-		DateTime,
-		Int,
-		Image,
-		Object,
-        Minutes
-	}
-}
