--- EmbyReports/Api/Common/ReportHeaderIdType.cs
@@ -1,8 +1,0 @@
-namespace EmbyReports.Api.Common
-{
-    public enum ReportHeaderIdType
-	{
-		Row,
-		Item
-	}
-}
