--- EmbyReports/Api/Common/ReportIncludeItemTypes.cs
@@ -1,20 +1,0 @@
-namespace EmbyReports.Api.Common
-{
-    public enum ReportIncludeItemTypes
-	{
-		MusicArtist,
-		MusicAlbum,
-		Book,
-		BoxSet,
-		Episode,
-		Video,
-		Movie,
-		MusicVideo,
-		Trailer,
-		Season,
-		Series,
-		Audio,
-		BaseItem,
-		Artist
-	}
-}
