--- EmbyReports/Api/Common/ReportViewType.cs
@@ -1,9 +1,0 @@
-namespace EmbyReports.Api.Common
-{
-    public enum ReportViewType
-	{
-        ReportData,
-        ReportActivities
-
-	}
-}
