--- EmbyReports/Configuration/PluginConfiguration.cs
@@ -1,8 +1,0 @@
-﻿using MediaBrowser.Model.Plugins;
-
-namespace EmbyReports.Configuration
-{
-    public class PluginConfiguration : BasePluginConfiguration
-    {
-    }
-}
