+++ JellyfinReports/Api/Common/HeaderActivitiesMetadata.cs
@@ -1,0 +1,16 @@
+namespace JellyfinReports.Api.Common
+{
+    public enum HeaderActivitiesMetadata
+	{
+        None,
+        Name,
+        Overview,
+        ShortOverview,
+        Type,
+        Date,
+        UserPrimaryImageTag,
+        Severity,
+        Item,
+        User
+	}
+}
