+++ JellyfinReports/Api/Common/HeaderMetadata.cs
@@ -1,0 +1,68 @@
+namespace JellyfinReports.Api.Common
+{
+    public enum HeaderMetadata
+	{
+		None,
+		Path,
+		Name,
+		PremiereDate,
+		DateAdded,
+		ReleaseDate,
+		Runtime,
+		PlayCount,
+		Season,
+		SeasonNumber,
+		Series,
+		Network,
+		Year,
+		ParentalRating,
+		CommunityRating,
+		Trailers,
+		Specials,
+		AlbumArtist,
+		Album,
+		Disc,
+		Track,
+		Audio,
+		EmbeddedImage,
+		Video,
+		Resolution,
+		Subtitles,
+		Genres,
+		Countries,
+        Status,
+		Tracks,
+		EpisodeSeries,
+		EpisodeSeason,
+        EpisodeNumber,
+        AudioAlbumArtist,
+		MusicArtist,
+		AudioAlbum,
+        Locked,
+        ImagePrimary,
+        ImageBackdrop,
+        ImageLogo,
+        Actor, 
+        Studios,
+        Composer, 
+        Director, 
+        GuestStar, 
+        Producer, 
+        Writer,
+        Artist,
+        Years,
+        ParentalRatings,
+        CommunityRatings,
+
+        //Activity logs
+        Overview,
+        ShortOverview,
+        Type,
+        Date,
+        UserPrimaryImage,
+        Severity,
+        Item,
+        User,
+        UserId
+	}
+}
