+++ JellyfinReports/Api/Common/ItemViewType.cs
@@ -1,0 +1,21 @@
+namespace JellyfinReports.Api.Common
+{
+    public enum ItemViewType
+    {
+        None,
+        Detail,
+        Edit,
+        List,
+        ItemByNameDetails,
+        StatusImage,
+        EmbeddedImage,
+        SubtitleImage,
+        TrailersImage,
+        SpecialsImage,
+        LockDataImage,
+        TagsPrimaryImage,
+        TagsBackdropImage,
+        TagsLogoImage,
+        UserPrimaryImage
+    }
+}
