+++ JellyfinReports/Api/Common/ReportDisplayType.cs
@@ -1,0 +1,10 @@
+namespace JellyfinReports.Api.Common
+{
+    public enum ReportDisplayType
+    {
+		None,
+		Screen,
+		Export,
+        ScreenExport 
+	}
+}
