+++ JellyfinReports/Api/Common/ReportExportType.cs
@@ -1,0 +1,8 @@
+namespace JellyfinReports.Api.Common
+{
+    public enum ReportExportType
+	{
+		CSV,
+		Excel
+	}
+}
