+++ JellyfinReports/Api/Common/ReportFieldType.cs
@@ -1,0 +1,15 @@
+namespace JellyfinReports.Api.Common
+{
+    public enum ReportFieldType
+	{
+		String,
+		Boolean,
+		Date,
+		Time,
+		DateTime,
+		Int,
+		Image,
+		Object,
+        Minutes
+	}
+}
