+++ JellyfinReports/Api/Common/ReportHeaderIdType.cs
@@ -1,0 +1,8 @@
+namespace JellyfinReports.Api.Common
+{
+    public enum ReportHeaderIdType
+	{
+		Row,
+		Item
+	}
+}
