+++ JellyfinReports/Api/Common/ReportIncludeItemTypes.cs
@@ -1,0 +1,20 @@
+namespace JellyfinReports.Api.Common
+{
+    public enum ReportIncludeItemTypes
+	{
+		MusicArtist,
+		MusicAlbum,
+		Book,
+		BoxSet,
+		Episode,
+		Video,
+		Movie,
+		MusicVideo,
+		Trailer,
+		Season,
+		Series,
+		Audio,
+		BaseItem,
+		Artist
+	}
+}
