+++ JellyfinReports/Api/Common/ReportViewType.cs
@@ -1,0 +1,9 @@
+namespace JellyfinReports.Api.Common
+{
+    public enum ReportViewType
+	{
+        ReportData,
+        ReportActivities
+
+	}
+}
