+++ JellyfinReports/Configuration/PluginConfiguration.cs
@@ -1,0 +1,8 @@
+﻿using MediaBrowser.Model.Plugins;
+
+namespace JellyfinReports.Configuration
+{
+    public class PluginConfiguration : BasePluginConfiguration
+    {
+    }
+}
